+++ aioamqp/channel.py
@@ -685,7 +685,8 @@ class Channel:
         if event:
             yield from event.wait()
             del self._ctag_events[consumer_tag]
-        yield from callback(body, envelope, properties)
+
+        yield from callback(self, body, envelope, properties)
 
     @asyncio.coroutine
     def server_basic_cancel(self, frame):